requirements.txt
# Core Web Framework Dependencies
flask>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=1.4.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Testing and Development
pytest>=6.0.0
pytest-asyncio>=0.21.0
python-dotenv>=0.19.0
click>=8.0.0

# Cloud Provider SDKs
google-cloud-security-center>=1.0.0
google-cloud-iam>=2.0.0
google-cloud-logging>=3.0.0
azure-identity>=1.10.0
azure-mgmt-security>=3.0.0
azure-mgmt-storage>=20.0.0
azure-mgmt-network>=20.0.0
boto3>=1.26.0

# Core Utilities
requests>=2.32.4
pyyaml>=6.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0

# Streamlit Dashboard Dependencies
streamlit>=1.46.0
plotly>=6.1.0
pandas>=2.3.0

# AI and Machine Learning
openai>=1.90.0
tiktoken>=0.9.0
numpy>=2.3.0
scikit-learn>=1.3.0
faiss-cpu>=1.11.0

# Vector Database and Semantic Search
weaviate-client>=4.15.0

# Security and Analysis
beautifulsoup4>=4.13.4
lxml>=5.4.0
cvss>=3.4
python-nmap>=0.7.1
stix2>=3.0.1

# Report Generation
reportlab>=4.4.2
markdown>=3.8.2

# Data Processing
trafilatura>=2.0.0
xmltodict>=0.14.2

# Google Coral TPU Dependencies (optional)
pycoral>=2.0.0
tflite-runtime>=2.13.0

# Redis for caching and task queues
redis>=4.0.0
celery>=5.3.0

# Monitoring and Observability
prometheus-client>=0.17.0
structlog>=23.1.0

# Security
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4